Rename address helpers file and change USA postal code label to "ZIP Code"
